cleaned up a bit, updated docs
